+++ internal/test/accessor/oneof_test.go
@@ -9,63 +9,64 @@ import (
 func TestOneOfAccessor(t *testing.T) {
 	t.Parallel()
 
-	t.Run("it returns the set value and ok as true to signify the presence of the value", func(t *testing.T) {
-		testAccessor(
-			t,
-			(*OneOf).SetFieldA,
-			(*OneOf).TryFieldA,
-			123,
-		)
+	t.Run("when the one-of value is set", func(t *testing.T) {
+		t.Run("it returns the value with ok set to true", func(t *testing.T) {
+			testAccessor(
+				t,
+				(*OneOf).SetFieldA,
+				(*OneOf).TryGetFieldA,
+				123,
+			)
 
-		testAccessor(
-			t,
-			(*OneOf).SetFieldA,
-			(*OneOf).TryFieldA,
-			0,
-		)
+			testAccessor(
+				t,
+				(*OneOf).SetFieldA,
+				(*OneOf).TryGetFieldA,
+				0,
+			)
 
-		testAccessor(
-			t,
-			(*OneOf).SetFieldB,
-			(*OneOf).TryFieldB,
-			456,
-		)
+			testAccessor(
+				t,
+				(*OneOf).SetFieldB,
+				(*OneOf).TryGetFieldB,
+				456,
+			)
 
-		testAccessor(
-			t,
-			(*OneOf).SetFieldB,
-			(*OneOf).TryFieldB,
-			0,
-		)
+			testAccessor(
+				t,
+				(*OneOf).SetFieldB,
+				(*OneOf).TryGetFieldB,
+				0,
+			)
 
-		testAccessor(
-			t,
-			(*OneOf).SetFieldC,
-			(*OneOf).TryFieldC,
-			"<value>",
-		)
+			testAccessor(
+				t,
+				(*OneOf).SetFieldC,
+				(*OneOf).TryGetFieldC,
+				"<value>",
+			)
 
-		testAccessor(
-			t,
-			(*OneOf).SetFieldC,
-			(*OneOf).TryFieldC,
-			"",
-		)
+			testAccessor(
+				t,
+				(*OneOf).SetFieldC,
+				(*OneOf).TryGetFieldC,
+				"",
+			)
+		})
 	})
 
-	t.Run("it ok as false to signify the absence of the value", func(t *testing.T) {
-		m := &OneOf{}
+	t.Run("when the one-of value is set to some other option", func(t *testing.T) {
+		t.Run("it returns with ok set to false", func(t *testing.T) {
+			m := &OneOf{}
+			m.SetFieldB(100)
 
-		if _, ok := m.TryFieldA(); ok {
-			t.Fatalf("TryFieldA() returned ok as true, want false")
-		}
+			if _, ok := m.TryGetFieldA(); ok {
+				t.Fatalf("TryGetFieldA() returned ok as true, want false")
+			}
 
-		if _, ok := m.TryFieldB(); ok {
-			t.Fatalf("TryFieldB() returned ok as true, want false")
-		}
-
-		if _, ok := m.TryFieldC(); ok {
-			t.Fatalf("TryFieldC() returned ok as true, want false")
-		}
+			if _, ok := m.TryGetFieldC(); ok {
+				t.Fatalf("TryGetFieldC() returned ok as true, want false")
+			}
+		})
 	})
 }